Add compliant test cases
